simple-kg/requirements.txt
neo4j
python-dotenv
beautifulsoup4